Move functions and objects for css style from LoginPage.js to SetLoginPage.js
